src/requirements.txt
boto3>=1.34.0
langchain-aws>=0.1.0
langchain-text-splitters>=0.2.0
xmltodict>=0.13.0
python-dotenv>=1.0.0